requirements.txt
django>=3.1.3
boto==2.49.0
boto3
# botocore==1.20.46
# moto==2.0.5
requests>=2.25.0
pytz>=2021.1
django-rename-app>=0.1.2
psycopg2-binary>=2.8.6
Pillow>=8.1.0
pyTelegramBotAPI>=3.7.7
djangorestframework>=3.12.2
PyJWT>=2.0.1
random_username
mailchecker
drf-nested-routers
opencv-python
pytesseract
numpy